Replace subscription service implementation
