+++ cmd/subscription/main.go
@@ -14,7 +14,10 @@ func main() {
 	cmd.GrpcServer(func(s *grpc.Server, db *gorm.DB, logger *zap.Logger) error {
 		api.RegisterSubscriptionServiceServer(
 			s,
-			service.NewSubscription(mysql.NewPersistentSubscription(db), logger),
+			service.NewSubscription(
+				mysql.NewStdPersistent(db, &mysql.SubscriptionModel{}),
+				logger,
+			),
 		)
 		return nil
 	})
