--- internal/mysql/standard.go
@@ -1,65 +1,0 @@
-package mysql
-
-import (
-	"errors"
-	"fmt"
-
-	"gorm.io/gorm"
-)
-
-func createMethod(model interface{}, db *gorm.DB) error {
-	err := db.Transaction(func(tx *gorm.DB) error {
-		return tx.Create(model).Error
-	})
-
-	if err != nil {
-		return fmt.Errorf("%w: %s", ErrUnknown, err)
-	}
-
-	return nil
-}
-
-func listMethod(models interface{}, db *gorm.DB, offset, limit int) error {
-	err := db.Transaction(func(tx *gorm.DB) error {
-		res := tx.Offset(offset).Limit(limit).Find(models)
-		if res.Error != nil {
-			return res.Error
-		}
-
-		return nil
-	})
-	if err != nil {
-		return fmt.Errorf("%w: %s", ErrUnknown, err)
-	}
-	return nil
-}
-
-func deleteMethod(model interface{}, db *gorm.DB, name string) error {
-	err := db.Transaction(func(tx *gorm.DB) error {
-		res := tx.Where("name = ?", name).Delete(model)
-		return res.Error
-	})
-
-	if err != nil {
-		return fmt.Errorf("%w: %s", ErrUnknown, err)
-	}
-
-	return nil
-}
-
-func getMethod(model interface{}, db *gorm.DB, name string) error {
-	err := db.Transaction(func(tx *gorm.DB) error {
-		res := tx.Where("name = ?", name).Take(model)
-		return res.Error
-	})
-
-	if errors.Is(err, gorm.ErrRecordNotFound) {
-		return fmt.Errorf("%w: %s", ErrNotFound, name)
-	}
-
-	if err != nil {
-		return fmt.Errorf("%w: %s", ErrUnknown, err)
-	}
-
-	return nil
-}
